[ FEATURE ] Ajuste no histórico de comandos para permitir refazer um comando e iniciar apenas uma instância por vez
