--- src/main/resources/static/typescript/application/history/commandHistory.ts
@@ -1,14 +1,0 @@
-import ICommand from "model/command/iCommand";
-
-export default class CommandHistory {
-  private _commands: ICommand[] = [];
-
-  public saveAndExecuteCommand(command: ICommand): void {
-    this._commands.push(command);
-    command.execute();
-  }
-
-  public undoLastCommand(): void {
-    this._commands.pop()?.undo();
-  }
-}
